`xtask`: Separate snapshot test functions
