+++ xtask/src/main.rs
@@ -1,60 +1,20 @@
-use std::{process::Command, str::FromStr, sync::Mutex};
-
-use anyhow::{anyhow, Context};
-use clap::{Parser, Subcommand};
-use colored::Colorize;
-use once_cell::sync::Lazy;
-use similar::{ChangeTag, TextDiff};
-
 mod backcompat;
+mod snapshot;
 mod targets;
 mod utils;
 
-use crate::utils::{
-    load_expected_output, overwrite_expected_output, run_capturing_stdout, run_command, rustc_is_nightly,
+use std::sync::Mutex;
+
+use anyhow::anyhow;
+use clap::{Parser, Subcommand};
+use once_cell::sync::Lazy;
+
+use crate::{
+    snapshot::{test_snapshot, Snapshot, ALL_SNAPSHOT_TESTS, SNAPSHOT_TESTS_DIRECTORY},
+    utils::{run_capturing_stdout, run_command},
 };
 
 static ALL_ERRORS: Lazy<Mutex<Vec<String>>> = Lazy::new(|| Mutex::new(vec![]));
-
-const SNAPSHOT_TESTS_DIRECTORY: &str = "firmware/qemu";
-const ALL_SNAPSHOT_TESTS: [&str; 12] = [
-    "log",
-    "bitflags",
-    "timestamp",
-    "panic",
-    "assert",
-    "assert-eq",
-    "assert-ne",
-    "unwrap",
-    "defmt-test",
-    "hints",
-    "hints_inner",
-    "dbg",
-];
-
-#[derive(Clone, Debug)]
-struct Snapshot(String);
-
-impl Snapshot {
-    pub fn name(&self) -> &str {
-        &self.0
-    }
-}
-
-impl FromStr for Snapshot {
-    type Err = String;
-
-    fn from_str(test: &str) -> Result<Self, Self::Err> {
-        if ALL_SNAPSHOT_TESTS.contains(&test) {
-            Ok(Self(String::from(test)))
-        } else {
-            Err(format!(
-                "Specified test '{}' does not exist, available tests are: {:?}",
-                test, ALL_SNAPSHOT_TESTS
-            ))
-        }
-    }
-}
 
 #[derive(Debug, Parser)]
 struct Options {
@@ -279,91 +239,6 @@ fn test_cross() {
         },
         "lint",
     );
-}
-
-fn test_snapshot(overwrite: bool, snapshot: Option<Snapshot>) {
-    println!("🧪 qemu/snapshot");
-
-    match snapshot {
-        None => test_all_snapshots(overwrite),
-        Some(snapshot) => {
-            do_test(
-                || test_single_snapshot(snapshot.name(), "", overwrite),
-                "qemu/snapshot",
-            );
-        }
-    }
-}
-
-fn test_all_snapshots(overwrite: bool) {
-    let mut tests = ALL_SNAPSHOT_TESTS.to_vec();
-
-    if rustc_is_nightly() {
-        tests.push("alloc");
-    }
-
-    for test in tests {
-        let features = if test == "alloc" { "alloc" } else { "" };
-
-        do_test(
-            || test_single_snapshot(test, features, overwrite),
-            "qemu/snapshot",
-        );
-    }
-}
-
-fn test_single_snapshot(name: &str, features: &str, overwrite: bool) -> anyhow::Result<()> {
-    println!("{}", name.bold());
-
-    let is_test = name.contains("test");
-
-    let mut args = if is_test {
-        vec!["-q", "tt", name]
-    } else {
-        vec!["-q", "rb", name]
-    };
-
-    if !features.is_empty() {
-        args.extend_from_slice(&["--features", features]);
-    }
-
-    let actual = run_capturing_stdout(
-        Command::new("cargo")
-            .args(&args)
-            .env("DEFMT_LOG", "trace")
-            .current_dir(SNAPSHOT_TESTS_DIRECTORY),
-    )
-    .with_context(|| name.to_string())?;
-
-    if overwrite {
-        overwrite_expected_output(name, actual.as_bytes(), is_test)?;
-        return Ok(());
-    }
-
-    let expected = load_expected_output(name, is_test)?;
-    let diff = TextDiff::from_lines(&expected, &actual);
-
-    // if anything isn't ChangeTag::Equal, print it and turn on error flag
-    let mut actual_matches_expected = true;
-    for op in diff.ops() {
-        for change in diff.iter_changes(op) {
-            let styled_change = match change.tag() {
-                ChangeTag::Delete => Some(("-".bold().red(), change.to_string().red())),
-                ChangeTag::Insert => Some(("+".bold().green(), change.to_string().green())),
-                ChangeTag::Equal => None,
-            };
-            if let Some((sign, change)) = styled_change {
-                actual_matches_expected = false;
-                eprint!("{sign}{change}");
-            }
-        }
-    }
-
-    if actual_matches_expected {
-        Ok(())
-    } else {
-        Err(anyhow!("{}", name))
-    }
 }
 
 fn test_book() {
